Move report related-code to new file
