+++ bench.go
@@ -4,9 +4,7 @@ import (
 	"bytes"
 	"crypto/rand"
 	"encoding/base32"
-	"encoding/json"
 	"fmt"
-	htmltemplate "html/template"
 	"log"
 	"os"
 	"os/exec"
@@ -15,14 +13,9 @@ import (
 	"strings"
 	"text/template"
 	"time"
-
-	vegeta "github.com/tsenart/vegeta/v12/lib"
 )
 
-var (
-	dockerComposeTemplate = template.Must(template.ParseFiles(filepath.Join("template", "docker-compose.yml.tmpl")))
-	summaryTemplate       = htmltemplate.Must(htmltemplate.ParseFiles(filepath.Join("template", "summary.html.tmpl")))
-)
+var dockerComposeTemplate = template.Must(template.ParseFiles(filepath.Join("template", "docker-compose.yml.tmpl")))
 
 type BenchmarkConfig struct {
 	ID        BenchmarkID
@@ -54,17 +47,6 @@ type RunConfig struct {
 	NeedsRelay bool
 }
 
-type SummaryFileData struct {
-	Name string
-	HDR  string
-	JSON TestResult
-}
-
-type SummaryFile struct {
-	Title string
-	Data  []SummaryFileData
-}
-
 type DockerComposeData struct {
 	ID         BenchmarkID
 	RunName    string
@@ -82,33 +64,6 @@ type App struct {
 }
 
 type BenchmarkID [4]byte
-
-type TestResult struct {
-	*vegeta.Metrics
-	Stats map[string]Stats `json:"container_stats"`
-}
-
-type Stats struct {
-	Before     ContainerStats           `json:"before"`
-	After      ContainerStats           `json:"after"`
-	Difference ContainerStatsDifference `json:"difference"`
-}
-
-type ContainerStats struct {
-	Timestamp           time.Time `json:"timestamp"`
-	MemoryMaxUsageBytes uint64    `json:"memory_max_usage_bytes"`
-	CPUUsageUser        uint64    `json:"cpu_usage_user"`
-	CPUUsageSystem      uint64    `json:"cpu_usage_system"`
-	CPUUsageTotal       uint64    `json:"cpu_usage_total"`
-}
-
-type ContainerStatsDifference struct {
-	Duration            time.Duration `json:"duration"`
-	MemoryMaxUsageBytes int64         `json:"memory_max_usage_bytes"`
-	CPUUsageUser        int64         `json:"cpu_usage_user"`
-	CPUUsageSystem      int64         `json:"cpu_usage_system"`
-	CPUUsageTotal       int64         `json:"cpu_usage_total"`
-}
 
 func NewBenchmarkID() BenchmarkID {
 	var id BenchmarkID
@@ -134,7 +89,7 @@ func Benchmark(cfg BenchmarkConfig) {
 		results = append(results, run(cfg, runCfg))
 	}
 
-	compare(results)
+	report(results)
 }
 
 type RunResult struct {
@@ -240,59 +195,5 @@ func waitUntilExit(containerName string) {
 
 	if status := string(bytes.TrimSpace(b)); status != "0" {
 		panic(fmt.Errorf("Container %s exited with status %s", containerName, status))
-	}
-}
-
-func read(path string) string {
-	file, err := os.ReadFile(path)
-	if err != nil {
-		panic(err)
-	}
-
-	return string(file)
-}
-
-func mustJSONUnmarshal(path string) (out TestResult) {
-	if err := json.Unmarshal([]byte(read(path)), &out); err != nil {
-		panic(err)
-	}
-	return out
-}
-
-func compare(results []*RunResult) {
-	summaryFile := SummaryFile{
-		Data: make([]SummaryFileData, 2),
-	}
-
-	for i, res := range results {
-		folderPath := filepath.Dir(res.Path)
-		name := res.Name
-
-		if i == 0 {
-			summaryFile.Title = folderPath
-		}
-		summaryFile.Data[i].Name = name
-		summaryFile.Data[i].HDR = read(filepath.Join(folderPath, name+".hdr"))
-		summaryFile.Data[i].JSON = mustJSONUnmarshal(filepath.Join(folderPath, name+".json"))
-	}
-
-	var b bytes.Buffer
-	err := summaryTemplate.Execute(&b, summaryFile)
-	if err != nil {
-		panic(err)
-	}
-
-	summaryPath := filepath.Join(summaryFile.Title, "summary.html")
-	fmt.Printf("Generating benchmark summary at %s", summaryPath)
-	if err := os.WriteFile(summaryPath, b.Bytes(), 0666); err != nil {
-		panic(err)
-	}
-
-	cmd := exec.Command(
-		"open",
-		summaryPath,
-	)
-	if err := cmd.Run(); err != nil {
-		panic(err)
 	}
 }
